Incorporated HW3 soln
