+++ memoryfs_shell_rpc.py
@@ -4,338 +4,273 @@ import argparse
 from memoryfs_client import *
 import os.path
 
-
 ## This class implements an interactive shell to navigate the file system
 
 class FSShell():
-    def __init__(self, file):
-        # cwd stored the inode of the current working directory
-        # we start in the root directory
-        self.cwd = 0
-        self.FileObject = file
+  def __init__(self, file):
+    # cwd stored the inode of the current working directory
+    # we start in the root directory
+    self.cwd = 0
+    self.FileObject = file
 
-    # implements cd (change directory)
-    def cd(self, dir):
-        # acquire lock
+  # implements cd (change directory)
+  def cd(self, dir):
+    # i = self.FileObject.Lookup(dir,self.cwd)
+    i = self.FileObject.GeneralPathToInodeNumber(dir,self.cwd)
+    if i == -1:
+      print ("Error: not found\n")
+      return -1
+    inobj = InodeNumber(self.FileObject.RawBlocks,i)
+    inobj.InodeNumberToInode()
+    if inobj.inode.type != INODE_TYPE_DIR:
+      print ("Error: not a directory\n")
+      return -1
+    self.cwd = i
+
+  # implements mkdir
+  def mkdir(self, dir):
+    i = self.FileObject.Create(self.cwd, dir, INODE_TYPE_DIR)
+    if i == -1:
+      print ("Error: cannot create directory\n")
+      return -1
+    return 0
+
+  # implements create
+  def create(self, file):
+    i = self.FileObject.Create(self.cwd, file, INODE_TYPE_FILE)
+    if i == -1:
+      print ("Error: cannot create file\n")
+      return -1
+    return 0
+
+  # implements append
+  def append(self, filename, string):
+    i = self.FileObject.Lookup(filename, self.cwd)
+    if i == -1:
+      print ("Error: not found\n")
+      return -1
+    inobj = InodeNumber(self.FileObject.RawBlocks,i)
+    inobj.InodeNumberToInode()
+    if inobj.inode.type != INODE_TYPE_FILE:
+      print ("Error: not a file\n")
+      return -1
+    written = self.FileObject.Write(i, inobj.inode.size, bytearray(string,"utf-8"))
+    print ("Successfully appended " + str(written) + " bytes.")
+    return 0
+    
+  # implements link
+  def link(self, target, name):
+    i = self.FileObject.Link(target, name, self.cwd)
+    if i == -1:
+      print ("Error: cannot create link\n")
+      return -1
+    return 0
+
+  # implements chroot
+  def chroot(self, target):
+    i = self.FileObject.Chroot(target, self.cwd)
+    if i == -1:
+      print ("Error: cannot chroot\n")
+      return -1
+    return 0
+
+
+  # implements ls (lists files in directory)
+  def ls(self):
+    inobj = InodeNumber(self.FileObject.RawBlocks, self.cwd)
+    inobj.InodeNumberToInode()
+    block_index = 0
+    while block_index <= (inobj.inode.size // BLOCK_SIZE):
+      block = self.FileObject.RawBlocks.Get(inobj.inode.block_numbers[block_index])
+      if block_index == (inobj.inode.size // BLOCK_SIZE):
+        end_position = inobj.inode.size % BLOCK_SIZE
+      else:
+        end_position = BLOCK_SIZE
+      current_position = 0
+      while current_position < end_position:
+        entryname = block[current_position:current_position+MAX_FILENAME]
+        entryinode = block[current_position+MAX_FILENAME:current_position+FILE_NAME_DIRENTRY_SIZE]
+        entryinodenumber = int.from_bytes(entryinode, byteorder='big')
+        inobj2 = InodeNumber(self.FileObject.RawBlocks, entryinodenumber)
+        inobj2.InodeNumberToInode()
+        if inobj2.inode.type == INODE_TYPE_DIR:
+          print (entryname.decode() + "/")
+        else:
+          print (entryname.decode())
+        current_position += FILE_NAME_DIRENTRY_SIZE
+      block_index += 1
+    return 0
+
+  # implements cat (print file contents)
+  def cat(self, filename):
+    i = self.FileObject.Lookup(filename, self.cwd)
+    if i == -1:
+      print ("Error: not found\n")
+      return -1
+    inobj = InodeNumber(self.FileObject.RawBlocks,i)
+    inobj.InodeNumberToInode()
+    if inobj.inode.type != INODE_TYPE_FILE:
+      print ("Error: not a file\n")
+      return -1
+    data = self.FileObject.Read(i, 0, MAX_FILE_SIZE)
+    print (data.decode())
+    return 0
+
+  # implements showblock (log block n contents)
+  def showblock(self, n):
+
+    try:
+      n = int(n)
+    except ValueError:
+      print('Error: ' + n + ' not a valid Integer')
+      return -1
+
+    if n < 0 or n >= TOTAL_NUM_BLOCKS:
+      print('Error: block number ' + str(n) + ' not in valid range [0, ' + str(TOTAL_NUM_BLOCKS - 1) + ']')
+      return -1
+    logging.info('Block (string) [' + str(n) + '] : ' + str((self.FileObject.RawBlocks.Get(n).decode(encoding='UTF-8',errors='ignore'))))
+    logging.info('Block (hex) [' + str(n) + '] : ' + str((self.FileObject.RawBlocks.Get(n).hex())))
+    return 0
+
+  # implements showinode (log inode i contents)
+  def showinode(self, i):
+
+    try:
+      i = int(i)
+    except ValueError:
+      print('Error: ' + i + ' not a valid Integer')
+      return -1
+
+    if i < 0 or i >= MAX_NUM_INODES:
+      print('Error: inode number ' + str(i) + ' not in valid range [0, ' + str(MAX_NUM_INODES - 1) + ']')
+      return -1
+
+    inobj = InodeNumber(self.FileObject.RawBlocks, i)
+    inobj.InodeNumberToInode()
+    inode = inobj.inode
+    inode.Print()
+    return 0
+
+  # implements showfsconfig (log fs config contents)
+  def showfsconfig(self):
+    self.FileObject.RawBlocks.PrintFSInfo()
+    return 0
+
+  # implements load (load the specified dump file)
+  def load(self, dumpfilename):
+    if not os.path.isfile(dumpfilename):
+      print("Error: Please provide valid file")
+      return -1
+    self.FileObject.RawBlocks.LoadFromDisk(dumpfilename)
+    self.cwd = 0
+    return 0
+
+  # implements save (save the file system contents to specified dump file)
+  def save(self, dumpfilename):
+    self.FileObject.RawBlocks.DumpToDisk(dumpfilename)
+    return 0
+
+  def Interpreter(self):
+    while (True):
+      command = input("[cwd=" + str(self.cwd) + "]:")
+      splitcmd = command.split()
+      if len(splitcmd) == 0:
+        continue
+      elif splitcmd[0] == "cd":
+        if len(splitcmd) != 2:
+          print ("Error: cd requires one argument")
+        else:
+          self.FileObject.RawBlocks.Acquire()
+          self.FileObject.RawBlocks.CheckAndInvalidate()
+          self.cd(splitcmd[1])
+          self.FileObject.RawBlocks.Release()
+      elif splitcmd[0] == "cat":
+        if len(splitcmd) != 2:
+          print ("Error: cat requires one argument")
+        else:
+          self.FileObject.RawBlocks.Acquire()
+          self.FileObject.RawBlocks.CheckAndInvalidate()
+          self.cat(splitcmd[1])
+          self.FileObject.RawBlocks.Release()
+      elif splitcmd[0] == "mkdir":
+        if len(splitcmd) != 2:
+          print ("Error: mkdir requires one argument")
+        else:
+          self.FileObject.RawBlocks.Acquire()
+          self.FileObject.RawBlocks.CheckAndInvalidate()
+          self.mkdir(splitcmd[1])
+          self.FileObject.RawBlocks.ForceInvalidate()
+          self.FileObject.RawBlocks.Release()
+      elif splitcmd[0] == "create":
+        if len(splitcmd) != 2:
+          print ("Error: create requires one argument")
+        else:
+          self.FileObject.RawBlocks.Acquire()
+          self.FileObject.RawBlocks.CheckAndInvalidate()
+          self.create(splitcmd[1])
+          self.FileObject.RawBlocks.ForceInvalidate()
+          self.FileObject.RawBlocks.Release()
+      elif splitcmd[0] == "ln":
+        if len(splitcmd) != 3:
+          print ("Error: ln requires two arguments")
+        else:
+          self.FileObject.RawBlocks.Acquire()
+          self.FileObject.RawBlocks.CheckAndInvalidate()
+          self.link(splitcmd[1], splitcmd[2])
+          self.FileObject.RawBlocks.ForceInvalidate()
+          self.FileObject.RawBlocks.Release()
+      elif splitcmd[0] == "chroot":
+        if len(splitcmd) != 2:
+          print ("Error: chroot requires one argument")
+        else:
+          self.FileObject.RawBlocks.Acquire()
+          self.FileObject.RawBlocks.CheckAndInvalidate()
+          self.chroot(splitcmd[1])
+          self.FileObject.RawBlocks.Release()
+      elif splitcmd[0] == "append":
+        if len(splitcmd) != 3:
+          print ("Error: append requires two arguments")
+        else:
+          self.FileObject.RawBlocks.Acquire()
+          self.FileObject.RawBlocks.CheckAndInvalidate()
+          self.append(splitcmd[1], splitcmd[2])
+          self.FileObject.RawBlocks.ForceInvalidate()
+          self.FileObject.RawBlocks.Release()
+      elif splitcmd[0] == "ls":
         self.FileObject.RawBlocks.Acquire()
-        # clear cache if not valid
-        self.FileObject.RawBlocks.CheckCacheValid()
-
-        # i = self.FileObject.Lookup(dir,self.cwd)
-        i = self.FileObject.GeneralPathToInodeNumber(dir, self.cwd)
-        if i == -1:
-            print("Error: not found\n")
-            # release lock
-            self.FileObject.RawBlocks.Release()
-            return -1
-        inobj = InodeNumber(self.FileObject.RawBlocks, i)
-        inobj.InodeNumberToInode()
-        if inobj.inode.type != INODE_TYPE_DIR:
-            print("Error: not a directory\n")
-            # release lock
-            self.FileObject.RawBlocks.Release()
-            return -1
-        self.cwd = i
-        # release lock
+        self.FileObject.RawBlocks.CheckAndInvalidate()
+        self.ls()
         self.FileObject.RawBlocks.Release()
+      elif splitcmd[0] == "showblock":
+        if len(splitcmd) != 2:
+          print ("Error: showblock requires one argument")
+        else:
+          self.showblock(splitcmd[1])
+      elif splitcmd[0] == "showinode":
+        if len(splitcmd) != 2:
+          print ("Error: showinode requires one argument")
+        else:
+          self.showinode(splitcmd[1])
+      elif splitcmd[0] == "showfsconfig":
+        if len(splitcmd) != 1:
+          print ("Error: showfsconfig do not require argument")
+        else:
+          self.showfsconfig()
+      elif splitcmd[0] == "load":
+        if len(splitcmd) != 2:
+          print ("Error: load requires 1 argument")
+        else:
+          self.load(splitcmd[1])
+      elif splitcmd[0] == "save":
+        if len(splitcmd) != 2:
+          print ("Error: save requires 1 argument")
+        else:
+          self.save(splitcmd[1])
+      elif splitcmd[0] == "exit":
+        return
+      else:
+        print ("command " + splitcmd[0] + " not valid.\n")
 
-    # implements mkdir
-    def mkdir(self, dir):
-        # acquire lock
-        self.FileObject.RawBlocks.Acquire()
-        # Invalidate caches
-        self.FileObject.RawBlocks.InvalidateCaches()
-        i = self.FileObject.Create(self.cwd, dir, INODE_TYPE_DIR)
-        # release lock
-        self.FileObject.RawBlocks.Release()
-        if i == -1:
-            print("Error: cannot create directory\n")
-            return -1
-        return 0
-
-    # implements create
-    def create(self, file):
-        # acquire lock
-        self.FileObject.RawBlocks.Acquire()
-        # Invalidate caches
-        self.FileObject.RawBlocks.InvalidateCaches()
-        i = self.FileObject.Create(self.cwd, file, INODE_TYPE_FILE)
-        # release lock
-        self.FileObject.RawBlocks.Release()
-        if i == -1:
-            print("Error: cannot create file\n")
-            return -1
-        return 0
-
-    # implements append
-    def append(self, filename, string):
-        # acquire lock
-        self.FileObject.RawBlocks.Acquire()
-        # sleep for testing
-        # time.sleep(3)
-        # Invalidate caches
-        self.FileObject.RawBlocks.InvalidateCaches()
-        i = self.FileObject.Lookup(filename, self.cwd)
-        if i == -1:
-            # release lock
-            self.FileObject.RawBlocks.Release()
-            print("Error: not found\n")
-            return -1
-        inobj = InodeNumber(self.FileObject.RawBlocks, i)
-        inobj.InodeNumberToInode()
-        if inobj.inode.type != INODE_TYPE_FILE:
-            # release lock
-            self.FileObject.RawBlocks.Release()
-            print("Error: not a file\n")
-            return -1
-        written = self.FileObject.Write(i, inobj.inode.size, bytearray(string, "utf-8"))
-        print("Successfully appended " + str(written) + " bytes.")
-        # release lock
-        self.FileObject.RawBlocks.Release()
-        return 0
-
-    # implements link
-    def link(self, target, name):
-        # acquire lock
-        self.FileObject.RawBlocks.Acquire()
-        # Invalidate caches
-        self.FileObject.RawBlocks.InvalidateCaches()
-        i = self.FileObject.Link(target, name, self.cwd)
-        # release lock
-        self.FileObject.RawBlocks.Release()
-        if i == -1:
-            print("Error: cannot create link\n")
-            return -1
-        return 0
-
-    # implements chroot
-    def chroot(self, target):
-        # acquire lock
-        self.FileObject.RawBlocks.Acquire()
-        # Invalidate caches
-        self.FileObject.RawBlocks.InvalidateCaches()
-        i = self.FileObject.Chroot(target, self.cwd)
-        # release lock
-        self.FileObject.RawBlocks.Release()
-        if i == -1:
-            print("Error: cannot chroot\n")
-            return -1
-        return 0
-
-    # implements ls (lists files in directory)
-    def ls(self):
-        # acquire lock
-        self.FileObject.RawBlocks.Acquire()
-        # clear cache if not valid
-        self.FileObject.RawBlocks.CheckCacheValid()
-        inobj = InodeNumber(self.FileObject.RawBlocks, self.cwd)
-        inobj.InodeNumberToInode()
-        block_index = 0
-        while block_index <= (inobj.inode.size // BLOCK_SIZE):
-            block = self.FileObject.RawBlocks.Get(inobj.inode.block_numbers[block_index])
-            if block_index == (inobj.inode.size // BLOCK_SIZE):
-                end_position = inobj.inode.size % BLOCK_SIZE
-            else:
-                end_position = BLOCK_SIZE
-            current_position = 0
-            while current_position < end_position:
-                entryname = block[current_position:current_position + MAX_FILENAME]
-                entryinode = block[current_position + MAX_FILENAME:current_position + FILE_NAME_DIRENTRY_SIZE]
-                entryinodenumber = int.from_bytes(entryinode, byteorder='big')
-                inobj2 = InodeNumber(self.FileObject.RawBlocks, entryinodenumber)
-                inobj2.InodeNumberToInode()
-                if inobj2.inode.type == INODE_TYPE_DIR:
-                    print(entryname.decode() + "/")
-                else:
-                    print(entryname.decode())
-                current_position += FILE_NAME_DIRENTRY_SIZE
-            block_index += 1
-        # release lock
-        self.FileObject.RawBlocks.Release()
-        return 0
-
-    # implements cat (print file contents)
-    def cat(self, filename):
-        # acquire lock
-        self.FileObject.RawBlocks.Acquire()
-        # clear cache if not valid
-        self.FileObject.RawBlocks.CheckCacheValid()
-        i = self.FileObject.Lookup(filename, self.cwd)
-        if i == -1:
-            # release lock
-            self.FileObject.RawBlocks.Release()
-            print("Error: not found\n")
-            return -1
-        inobj = InodeNumber(self.FileObject.RawBlocks, i)
-        inobj.InodeNumberToInode()
-        if inobj.inode.type != INODE_TYPE_FILE:
-            # release lock
-            self.FileObject.RawBlocks.Release()
-            print("Error: not a file\n")
-            return -1
-        data = self.FileObject.Read(i, 0, MAX_FILE_SIZE)
-        print(data.decode())
-        # release lock
-        self.FileObject.RawBlocks.Release()
-        return 0
-
-    # implements showblock (log block n contents)
-    def showblock(self, n):
-        # acquire lock
-        self.FileObject.RawBlocks.Acquire()
-        # clear cache if not valid
-        self.FileObject.RawBlocks.CheckCacheValid()
-        try:
-            n = int(n)
-        except ValueError:
-            # release lock
-            self.FileObject.RawBlocks.Release()
-            print('Error: ' + n + ' not a valid Integer')
-            return -1
-
-        if n < 0 or n >= TOTAL_NUM_BLOCKS:
-            # release lock
-            self.FileObject.RawBlocks.Release()
-            print('Error: block number ' + str(n) + ' not in valid range [0, ' + str(TOTAL_NUM_BLOCKS - 1) + ']')
-            return -1
-        logging.info('Block (string) [' + str(n) + '] : ' + str(
-            (self.FileObject.RawBlocks.Get(n).decode(encoding='UTF-8', errors='ignore'))))
-        logging.info('Block (hex) [' + str(n) + '] : ' + str((self.FileObject.RawBlocks.Get(n).hex())))
-        # release lock
-        self.FileObject.RawBlocks.Release()
-        return 0
-
-    # implements showinode (log inode i contents)
-    def showinode(self, i):
-        # acquire lock
-        self.FileObject.RawBlocks.Acquire()
-        # clear cache if not valid
-        self.FileObject.RawBlocks.CheckCacheValid()
-        try:
-            i = int(i)
-        except ValueError:
-            # release lock
-            self.FileObject.RawBlocks.Release()
-            print('Error: ' + i + ' not a valid Integer')
-            return -1
-
-        if i < 0 or i >= MAX_NUM_INODES:
-            # release lock
-            self.FileObject.RawBlocks.Release()
-            print('Error: inode number ' + str(i) + ' not in valid range [0, ' + str(MAX_NUM_INODES - 1) + ']')
-            return -1
-
-        inobj = InodeNumber(self.FileObject.RawBlocks, i)
-        inobj.InodeNumberToInode()
-        inode = inobj.inode
-        inode.Print()
-        # release lock
-        self.FileObject.RawBlocks.Release()
-        return 0
-
-    # implements showfsconfig (log fs config contents)
-    def showfsconfig(self):
-        # acquire lock
-        self.FileObject.RawBlocks.Acquire()
-        # clear cache if not valid
-        self.FileObject.RawBlocks.CheckCacheValid()
-        self.FileObject.RawBlocks.PrintFSInfo()
-        # release lock
-        self.FileObject.RawBlocks.Release()
-        return 0
-
-    # implements load (load the specified dump file)
-    def load(self, dumpfilename):
-        # acquire lock
-        self.FileObject.RawBlocks.Acquire()
-        if not os.path.isfile(dumpfilename):
-            # release lock
-            self.FileObject.RawBlocks.Release()
-            print("Error: Please provide valid file")
-            return -1
-        self.FileObject.RawBlocks.LoadFromDisk(dumpfilename)
-        self.cwd = 0
-        # release lock
-        self.FileObject.RawBlocks.Release()
-        return 0
-
-    # implements save (save the file system contents to specified dump file)
-    def save(self, dumpfilename):
-        self.FileObject.RawBlocks.DumpToDisk(dumpfilename)
-        return 0
-
-    def Interpreter(self):
-        while (True):
-            command = input("[cwd=" + str(self.cwd) + "]:")
-            splitcmd = command.split()
-            if len(splitcmd) == 0:
-                continue
-            elif splitcmd[0] == "cd":
-                if len(splitcmd) != 2:
-                    print("Error: cd requires one argument")
-                else:
-                    self.cd(splitcmd[1])
-            elif splitcmd[0] == "cat":
-                if len(splitcmd) != 2:
-                    print("Error: cat requires one argument")
-                else:
-                    self.cat(splitcmd[1])
-            elif splitcmd[0] == "mkdir":
-                if len(splitcmd) != 2:
-                    print("Error: mkdir requires one argument")
-                else:
-                    self.mkdir(splitcmd[1])
-            elif splitcmd[0] == "create":
-                if len(splitcmd) != 2:
-                    print("Error: create requires one argument")
-                else:
-                    self.create(splitcmd[1])
-            elif splitcmd[0] == "ln":
-                if len(splitcmd) != 3:
-                    print("Error: ln requires two arguments")
-                else:
-                    self.link(splitcmd[1], splitcmd[2])
-            elif splitcmd[0] == "chroot":
-                if len(splitcmd) != 2:
-                    print("Error: chroot requires one argument")
-                else:
-                    self.chroot(splitcmd[1])
-            elif splitcmd[0] == "append":
-                if len(splitcmd) != 3:
-                    print("Error: append requires two arguments")
-                else:
-                    self.append(splitcmd[1], splitcmd[2])
-            elif splitcmd[0] == "ls":
-                self.ls()
-            elif splitcmd[0] == "showblock":
-                if len(splitcmd) != 2:
-                    print("Error: showblock requires one argument")
-                else:
-                    self.showblock(splitcmd[1])
-            elif splitcmd[0] == "showinode":
-                if len(splitcmd) != 2:
-                    print("Error: showinode requires one argument")
-                else:
-                    self.showinode(splitcmd[1])
-            elif splitcmd[0] == "showfsconfig":
-                if len(splitcmd) != 1:
-                    print("Error: showfsconfig do not require argument")
-                else:
-                    self.showfsconfig()
-            elif splitcmd[0] == "load":
-                if len(splitcmd) != 2:
-                    print("Error: load requires 1 argument")
-                else:
-                    self.load(splitcmd[1])
-            elif splitcmd[0] == "save":
-                if len(splitcmd) != 2:
-                    print("Error: save requires 1 argument")
-                else:
-                    self.save(splitcmd[1])
-            elif splitcmd[0] == "exit":
-                return
-            else:
-                print("command " + splitcmd[0] + " not valid.\n")
 
 
 if __name__ == "__main__":
